--- crates/opsml_sql/src/queries/scripts/get_version.sql
@@ -1,15 +1,0 @@
-SELECT
-    date,
-    timestamp,
-    name,
-    repository,
-    major,
-    minor,
-    patch,
-    pre_release,
-    build,
-    contact
-FROM $1
-WHERE 
-    AND name = $2
-    AND repository = $3
