+++ crates/opsml_sql/src/sqlite/mod.rs
@@ -1,3 +1,2 @@
 pub mod client;
-pub mod helper;
 pub mod schema;
